+++ lib/src/modules/tabby/core/models/tabby_credentials.dart
@@ -1,3 +1,5 @@
+import 'package:uni_pay/uni_pay.dart';
+
 class TabbyCredential {
   /// Tabby `publishable secret key` provided by Tabby to process the payments
   late String psKey;
@@ -8,12 +10,16 @@ class TabbyCredential {
   /// Tabby `secret key` provided by Tabby to process the payments, only needed for `get`, `capture`, and `refund` operations
   late String secretKey;
 
+  ///* Merchant `Urls` for Webhook notification
+  late MerchantUrl merchantUrl;
+
   /// Tabby credentials to process the payments
   TabbyCredential({
     required this.psKey,
     this.merchantCode = "sa",
     this.secretKey = "",
-  });
+    MerchantUrl? merchantUrl,
+  }) : merchantUrl = merchantUrl ?? MerchantUrl(notification: "");
 
   /// Convert the TabbyCredential to JSON
   Map<String, dynamic> toJson() => {